+++ src/main/java/com/qubole/qds/sdk/java/entities/SchemaList.java
@@ -15,47 +15,52 @@
  */
 package com.qubole.qds.sdk.java.entities;
 
-import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 import java.util.List;
-import java.util.Map;
 
-@JsonIgnoreProperties(ignoreUnknown = true)
 public class SchemaList {
-  private PagingInfo paging_info;
-  private Map<String, List<Schema>> schemas;
+
+  private List<String> schemas;
+  private String default_schema;
+  private String status;
 
   public SchemaList()
   {
   }
 
-  public SchemaList(Map<String, List<Schema>> schemas)
+  public SchemaList(List<String> schemas, String default_schema, String status)
   {
     this.schemas = schemas;
+    this.default_schema = default_schema;
+    this.status = status;
   }
 
-  public SchemaList(PagingInfo paging_info, Map<String, List<Schema>> schemas)
-  {
-    this.paging_info = paging_info;
-    this.schemas = schemas;
-  }
-
-  public PagingInfo getPaging_info()
-  {
-    return paging_info;
-  }
-
-  public void setPaging_info(PagingInfo paging_info)
-  {
-    this.paging_info = paging_info;
-  }
-
-  public Map<String, List<Schema>> getSchemas()
+  public List<String> getSchemas()
   {
     return schemas;
   }
 
-  public void setSchemas(Map<String, List<Schema>> schemas)
+  public String getDefault_schema()
+  {
+    return default_schema;
+  }
+
+  public String getStatus()
+  {
+    return status;
+  }
+
+  public void setSchemas(List<String> schemas)
   {
     this.schemas = schemas;
+  }
+
+  public void setDefault_schema(String default_schema)
+  {
+    this.default_schema = default_schema;
+  }
+
+  public void setStatus(String status)
+  {
+    this.status = status;
   }
 }
